created spawnCommandSync
